del print, correct main
