+++ src/vacancies_filter.py
@@ -64,25 +64,3 @@ class VacancyFilter:
     def sort_by_salary(self, direction=False):
         """Метод сортирует список объектов вакансий по зарплате (по-умолчанию по возрастанию)"""
         self.__vacs.sort(key=lambda x: x.salary, reverse=direction)
-
-
-if __name__ == "__main__":
-    # vac_1 = Vacancy("Менеджер", "url", "Продажа оборудования", "Без опыта работы.", "Москва", None)
-    # vac_2 = Vacancy("Менеджер", "url", "Разработчик Python", "Без опыта работы.", "Москва", 30000)
-    # v_filter = VacancyFilter()
-    # data = [vac_1, vac_2]
-    # v_filter.vacs = data
-    # print(*v_filter.vacs)
-    # # print(data)
-    # v_filter.filter_requirement('Менеджер')
-    # print(v_filter.vacs)
-    # v_filter.filter_description('Продажа')
-    # # print(v_filter.vacs)
-    # # v_filter.filter_area('Москва')
-    # # v_filter.filter_salary('30000')
-
-    hh = HH()
-    # a = hh.export_vac_list()
-    # hh.load_vacancies('менеджер')
-    file_v = VacancyFile()
-    file_v.import_vacancy_list(hh.vacancies)
